+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - refactor centralized CSS into components
 - fix git ignore
 - remove unused assets
+- implement dark/light theme toggle
 
 ## [0.1.1] - 2025-10-21